Added 4 new files for header, imdb, metacritic and rotten tomatoes
