+++ headers.py
@@ -1,0 +1,6 @@
+class headers_class():
+		def headers():
+        ua_one = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_10_1) '
+        ua_two = 'AppleWebKit/537.36 (KHTML, like Gecko) Chrome/39.0.2171.95 Safari/537.36'
+        headers = {'User-Agent': ua_one + ua_two}
+        return headers
